Ajustes generales y validaciones de solicitud de credito

+++ Fuentes/ManejoFondo/ManejoFondo/Entities/FondoAhorroMensualEntity.cs
@@ -15,7 +15,7 @@ namespace ManejoFondo.Entities
     {
         public int N_Id { get; set; }
         public long N_Id_Usuario { get; set; }
-        public double N_Valor_Cuota { get; set; }
+        public decimal N_Valor_Cuota { get; set; }
         public DateTime F_Fecha_Registro { get; set; }
         public int N_Mes_Ahorro { get; set; }
         public int N_Anio_Ahorro { get; set; }
